+++ opsdroid/conftest.py
@@ -3,4 +3,6 @@ from opsdroid.testing import opsdroid
 
 from opsdroid.cli.start import configure_lang
 
+__all__ = ["opsdroid"]
+
 configure_lang({})
