+++ setup.cfg
@@ -13,7 +13,7 @@ select = C,E,F,W,B,B950
 ignore = E203,E501,W503,F821
 
 [pydocstyle]
-ignore = D202,D203,D212,D213,D406,D407,F821
+ignore = D202,D203,D212,D213,D406,D407
 
 [versioneer]
 VCS = git
