+++ tox.ini
@@ -42,12 +42,13 @@ commands =
 
 [testenv:docs]
 basepython = python3
+changedir = docs
 ignore_errors = True
 whitelist_externals = sh
 commands =
-     sphinx-build -W -b html docs/ docs/_build/
+     sphinx-build --color -W --keep-going -b html . _build/ {posargs}
      ; running python's http server to serve documentation on 127.0.0.1:3001
-     sh -c 'cd docs/_build && python -m http.server 3001 2>&1 > /dev/null &'
+     sh -c 'cd _build/ && python -m http.server 3001 2>&1 > /dev/null &'
      ; checking liveness
      deadlinks http://127.0.0.1:3001/ -n 10 -r 3 --no-progress --fiff
